+++ scripts/homebrew/xvn.rb
@@ -11,22 +11,22 @@
 class Xvn < Formula
   desc "Self-hosted trading strategy engine with dashboard and optimizer"
   homepage "https://github.com/latentwill/xvision"
-  version "0.37.0"
+  version "0.38.0"
   license "Apache-2.0"
 
   on_macos do
     if Hardware::CPU.arm?
-      url "https://github.com/latentwill/xvision/releases/download/v0.37.0/xvn-aarch64-apple-darwin.tar.gz"
+      url "https://github.com/latentwill/xvision/releases/download/v0.38.0/xvn-aarch64-apple-darwin.tar.gz"
       sha256 "REPLACE_WITH_ACTUAL_SHA256_ARM64"
     else
-      url "https://github.com/latentwill/xvision/releases/download/v0.37.0/xvn-x86_64-apple-darwin.tar.gz"
+      url "https://github.com/latentwill/xvision/releases/download/v0.38.0/xvn-x86_64-apple-darwin.tar.gz"
       sha256 "REPLACE_WITH_ACTUAL_SHA256_X86_64"
     end
   end
 
   on_linux do
     if Hardware::CPU.intel?
-      url "https://github.com/latentwill/xvision/releases/download/v0.37.0/xvn-x86_64-linux-musl.tar.gz"
+      url "https://github.com/latentwill/xvision/releases/download/v0.38.0/xvn-x86_64-linux-musl.tar.gz"
       sha256 "REPLACE_WITH_ACTUAL_SHA256_LINUX"
     end
   end